renamed container folder files acording to the naming guidelines
