+++ src/client/App.js
@@ -10,14 +10,14 @@ import { useAuthentication } from './hooks/useAuthentication';
 import Profile from './containers/Profile';
 import Loader from './components/Loader';
 import Error404Page from './containers/404Page/404Page.container';
-import CareerPage from './containers/CareerPage/CareerPage.component';
-import { QuizPage } from './containers/QuizPage/QuizPage.component';
-import { QuizResultPage } from './containers/QuizResultPage/QuizResultPage.component';
+import CareerPage from './containers/CareerPage/CareerPage.container';
+import { QuizPage } from './containers/QuizPage/QuizPage.container';
+import { QuizResultPage } from './containers/QuizResultPage/QuizResultPage.container';
 import { EmptyQuizResultPage } from './containers/EmptyQuizResultPage/EmptyQuizResultPage.container';
-import { ElevatorPitchPage } from './containers/ElevatorPitchPage/ElevatorPitchPage.component';
-import { LoginPage } from './containers/LoginPage/LoginPage.component';
-import { NextStepsPage } from './containers/NextStepsPage/NextStepsPage.component';
-import RegistrationPage from './containers/RegistrationPage/RegistrationPage.component';
+import { ElevatorPitchPage } from './containers/ElevatorPitchPage/ElevatorPitchPage.container';
+import { LoginPage } from './containers/LoginPage/LoginPage.container';
+import { NextStepsPage } from './containers/NextStepsPage/NextStepsPage.container';
+import RegistrationPage from './containers/RegistrationPage/RegistrationPage.container';
 import ProfileBlueBar from './components/ProfileBlueBar/ProfileBlueBar.component';
 
 function App() {
