+++ src/client/containers/RegistrationPage/RegistrationPage.container.js
