+++ src/client/containers/RegistrationPage/RegistrationPage.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegistrationPage from './RegistrationPage.component.js';
+import RegistrationPage from './RegistrationPage.container';
 
 export default {
   title: 'containers / RegistrationPage ',
